Lab_3/requirements.txt
fastapi
uvicorn
sqlalchemy
python-jose[cryptography]
passlib[bcrypt]
cachetools
stomp.py==8.1.0